core: create a gsl_simulation app
